refactor: move yaml implementation to stdlib
