+++ crates/jrsonnet-evaluator/src/manifest.rs
@@ -1,12 +1,39 @@
 use std::{borrow::Cow, fmt::Write};
 
 use crate::{
-	error::{Error::*, Result},
-	throw, ManifestFormat, State, Val,
+	error::{ErrorKind::*, Result},
+	throw, State, Val,
 };
 
+pub trait ManifestFormat {
+	fn manifest_buf(&self, val: Val, buf: &mut String) -> Result<()>;
+	fn manifest(&self, val: Val) -> Result<String> {
+		let mut out = String::new();
+		self.manifest_buf(val, &mut out)?;
+		Ok(out)
+	}
+}
+impl<T> ManifestFormat for Box<T>
+where
+	T: ManifestFormat + ?Sized,
+{
+	fn manifest_buf(&self, val: Val, buf: &mut String) -> Result<()> {
+		let inner = &**self;
+		inner.manifest_buf(val, buf)
+	}
+}
+impl<T> ManifestFormat for &'_ T
+where
+	T: ManifestFormat + ?Sized,
+{
+	fn manifest_buf(&self, val: Val, buf: &mut String) -> Result<()> {
+		let inner = &**self;
+		inner.manifest_buf(val, buf)
+	}
+}
+
 #[derive(PartialEq, Eq, Clone, Copy)]
-pub enum ManifestType {
+enum JsonFormatting {
 	// Applied in manifestification
 	Manifest,
 	/// Used for std.manifestJson
@@ -20,7 +47,7 @@ pub enum ManifestType {
 
 pub struct JsonFormat<'s> {
 	padding: Cow<'s, str>,
-	mtype: ManifestType,
+	mtype: JsonFormatting,
 	newline: &'s str,
 	key_val_sep: &'s str,
 	#[cfg(feature = "exp-preserve-order")]
@@ -32,7 +59,7 @@ impl<'s> JsonFormat<'s> {
 	pub fn minify(#[cfg(feature = "exp-preserve-order")] preserve_order: bool) -> Self {
 		Self {
 			padding: Cow::Borrowed(""),
-			mtype: ManifestType::Minify,
+			mtype: JsonFormatting::Minify,
 			newline: "\n",
 			key_val_sep: ":",
 			#[cfg(feature = "exp-preserve-order")]
@@ -43,7 +70,7 @@ impl<'s> JsonFormat<'s> {
 	pub fn std_to_string() -> Self {
 		Self {
 			padding: Cow::Borrowed(""),
-			mtype: ManifestType::ToString,
+			mtype: JsonFormatting::ToString,
 			newline: "\n",
 			key_val_sep: ": ",
 			#[cfg(feature = "exp-preserve-order")]
@@ -58,7 +85,7 @@ impl<'s> JsonFormat<'s> {
 	) -> Self {
 		Self {
 			padding: Cow::Owned(padding),
-			mtype: ManifestType::Std,
+			mtype: JsonFormatting::Std,
 			newline,
 			key_val_sep,
 			#[cfg(feature = "exp-preserve-order")]
@@ -78,7 +105,7 @@ impl<'s> JsonFormat<'s> {
 		}
 		Self {
 			padding: Cow::Owned(" ".repeat(padding)),
-			mtype: ManifestType::Manifest,
+			mtype: JsonFormatting::Manifest,
 			newline: "\n",
 			key_val_sep: ": ",
 			#[cfg(feature = "exp-preserve-order")]
@@ -90,7 +117,7 @@ impl Default for JsonFormat<'static> {
 	fn default() -> Self {
 		Self {
 			padding: Cow::Borrowed("    "),
-			mtype: ManifestType::Manifest,
+			mtype: JsonFormatting::Manifest,
 			newline: "\n",
 			key_val_sep: ": ",
 			#[cfg(feature = "exp-preserve-order")]
@@ -125,7 +152,7 @@ fn manifest_json_ex_buf(
 		Val::Arr(items) => {
 			buf.push('[');
 			if !items.is_empty() {
-				if mtype != ManifestType::ToString && mtype != ManifestType::Minify {
+				if mtype != JsonFormatting::ToString && mtype != JsonFormatting::Minify {
 					buf.push_str(options.newline);
 				}
 
@@ -134,9 +161,9 @@ fn manifest_json_ex_buf(
 				for (i, item) in items.iter().enumerate() {
 					if i != 0 {
 						buf.push(',');
-						if mtype == ManifestType::ToString {
+						if mtype == JsonFormatting::ToString {
 							buf.push(' ');
-						} else if mtype != ManifestType::Minify {
+						} else if mtype != JsonFormatting::Minify {
 							buf.push_str(options.newline);
 						}
 					}
@@ -145,14 +172,14 @@ fn manifest_json_ex_buf(
 				}
 				cur_padding.truncate(old_len);
 
-				if mtype != ManifestType::ToString && mtype != ManifestType::Minify {
+				if mtype != JsonFormatting::ToString && mtype != JsonFormatting::Minify {
 					buf.push_str(options.newline);
 					buf.push_str(cur_padding);
 				}
-			} else if mtype == ManifestType::Std {
+			} else if mtype == JsonFormatting::Std {
 				buf.push_str("\n\n");
 				buf.push_str(cur_padding);
-			} else if mtype == ManifestType::ToString || mtype == ManifestType::Manifest {
+			} else if mtype == JsonFormatting::ToString || mtype == JsonFormatting::Manifest {
 				buf.push(' ');
 			}
 			buf.push(']');
@@ -165,7 +192,7 @@ fn manifest_json_ex_buf(
 				options.preserve_order,
 			);
 			if !fields.is_empty() {
-				if mtype != ManifestType::ToString && mtype != ManifestType::Minify {
+				if mtype != JsonFormatting::ToString && mtype != JsonFormatting::Minify {
 					buf.push_str(options.newline);
 				}
 
@@ -174,9 +201,9 @@ fn manifest_json_ex_buf(
 				for (i, field) in fields.into_iter().enumerate() {
 					if i != 0 {
 						buf.push(',');
-						if mtype == ManifestType::ToString {
+						if mtype == JsonFormatting::ToString {
 							buf.push(' ');
-						} else if mtype != ManifestType::Minify {
+						} else if mtype != JsonFormatting::Minify {
 							buf.push_str(options.newline);
 						}
 					}
@@ -194,14 +221,14 @@ fn manifest_json_ex_buf(
 				}
 				cur_padding.truncate(old_len);
 
-				if mtype != ManifestType::ToString && mtype != ManifestType::Minify {
+				if mtype != JsonFormatting::ToString && mtype != JsonFormatting::Minify {
 					buf.push_str(options.newline);
 					buf.push_str(cur_padding);
 				}
-			} else if mtype == ManifestType::Std {
+			} else if mtype == JsonFormatting::Std {
 				buf.push_str("\n\n");
 				buf.push_str(cur_padding);
-			} else if mtype == ManifestType::ToString || mtype == ManifestType::Manifest {
+			} else if mtype == JsonFormatting::ToString || mtype == JsonFormatting::Manifest {
 				buf.push(' ');
 			}
 			buf.push('}');
@@ -213,7 +240,7 @@ fn manifest_json_ex_buf(
 
 impl ManifestFormat for JsonFormat<'_> {
 	fn manifest_buf(&self, val: Val, buf: &mut String) -> Result<()> {
-		manifest_json_ex_buf(&val, buf, &mut String::new(), &self)
+		manifest_json_ex_buf(&val, buf, &mut String::new(), self)
 	}
 }
 
@@ -293,9 +320,9 @@ static ESCAPE: [u8; 256] = [
 	__, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // F
 ];
 
-fn escape_string_json_buf(value: &str, buf: &mut String) {
+pub fn escape_string_json_buf(value: &str, buf: &mut String) {
 	// Safety: we only write correct utf-8 in this function
-	let mut buf: &mut Vec<u8> = unsafe { core::mem::transmute(buf) };
+	let buf: &mut Vec<u8> = unsafe { &mut *(buf as *mut String).cast::<Vec<u8>>() };
 	let bytes = value.as_bytes();
 
 	// Perfect for ascii strings, removes any reallocations
@@ -318,7 +345,7 @@ fn escape_string_json_buf(value: &str, buf: &mut String) {
 
 		match escape {
 			self::BB | self::TT | self::NN | self::FF | self::RR | self::QU | self::BS => {
-				buf.extend_from_slice(&[b'\\', escape])
+				buf.extend_from_slice(&[b'\\', escape]);
 			}
 			self::UU => {
 				static HEX_DIGITS: [u8; 16] = *b"0123456789abcdef";
@@ -330,7 +357,7 @@ fn escape_string_json_buf(value: &str, buf: &mut String) {
 					HEX_DIGITS[(byte >> 4) as usize],
 					HEX_DIGITS[(byte & 0xF) as usize],
 				];
-				buf.extend_from_slice(bytes)
+				buf.extend_from_slice(bytes);
 			}
 			_ => unreachable!(),
 		}
@@ -343,215 +370,4 @@ fn escape_string_json_buf(value: &str, buf: &mut String) {
 
 	buf.extend_from_slice(&bytes[start..]);
 	buf.push(b'"');
-}
-
-pub struct YamlFormat<'s> {
-	/// Padding before fields, i.e
-	/// ```yaml
-	/// a:
-	///   b:
-	/// ## <- this
-	/// ```
-	padding: Cow<'s, str>,
-	/// Padding before array elements in objects
-	/// ```yaml
-	/// a:
-	///   - 1
-	/// ## <- this
-	/// ```
-	arr_element_padding: Cow<'s, str>,
-	/// Should yaml keys appear unescaped, when possible
-	/// ```yaml
-	/// "safe_key": 1
-	/// # vs
-	/// safe_key: 1
-	/// ```
-	quote_keys: bool,
-	/// If true - then order of fields is preserved as written,
-	/// instead of sorting alphabetically
-	#[cfg(feature = "exp-preserve-order")]
-	preserve_order: bool,
-}
-impl YamlFormat<'_> {
-	pub fn cli(
-		padding: usize,
-		#[cfg(feature = "exp-preserve-order")] preserve_order: bool,
-	) -> Self {
-		let padding = " ".repeat(padding);
-		Self {
-			padding: Cow::Owned(padding.clone()),
-			arr_element_padding: Cow::Owned(padding),
-			quote_keys: false,
-			#[cfg(feature = "exp-preserve-order")]
-			preserve_order,
-		}
-	}
-	pub fn std_to_yaml(
-		indent_array_in_object: bool,
-		quote_keys: bool,
-		#[cfg(feature = "exp-preserve-order")] preserve_order: bool,
-	) -> Self {
-		Self {
-			padding: Cow::Borrowed("  "),
-			arr_element_padding: Cow::Borrowed(if indent_array_in_object { "  " } else { "" }),
-			quote_keys,
-			#[cfg(feature = "exp-preserve-order")]
-			preserve_order,
-		}
-	}
-}
-impl ManifestFormat for YamlFormat<'_> {
-	fn manifest_buf(&self, val: Val, buf: &mut String) -> Result<()> {
-		manifest_yaml_ex_buf(&val, buf, &mut String::new(), self)
-	}
-}
-
-/// From <https://github.com/chyh1990/yaml-rust/blob/da52a68615f2ecdd6b7e4567019f280c433c1521/src/emitter.rs#L289>
-/// With added date check
-fn yaml_needs_quotes(string: &str) -> bool {
-	fn need_quotes_spaces(string: &str) -> bool {
-		string.starts_with(' ') || string.ends_with(' ')
-	}
-
-	string.is_empty()
-		|| need_quotes_spaces(string)
-		|| string.starts_with(|c| matches!(c, '&' | '*' | '?' | '|' | '-' | '<' | '>' | '=' | '!' | '%' | '@'))
-		|| string.contains(|c| matches!(c, ':' | '{' | '}' | '[' | ']' | ',' | '#' | '`' | '\"' | '\'' | '\\' | '\0'..='\x06' | '\t' | '\n' | '\r' | '\x0e'..='\x1a' | '\x1c'..='\x1f'))
-		|| [
-			// http://yaml.org/type/bool.html
-			// Note: 'y', 'Y', 'n', 'N', is not quoted deliberately, as in libyaml. PyYAML also parse
-			// them as string, not booleans, although it is violating the YAML 1.1 specification.
-			// See https://github.com/dtolnay/serde-yaml/pull/83#discussion_r152628088.
-			"yes", "Yes", "YES", "no", "No", "NO", "True", "TRUE", "true", "False", "FALSE", "false",
-			"on", "On", "ON", "off", "Off", "OFF", // http://yaml.org/type/null.html
-			"null", "Null", "NULL", "~",
-		].contains(&string)
-		|| (string.chars().all(|c| matches!(c, '0'..='9' | '-'))
-			&& string.chars().filter(|c| *c == '-').count() == 2)
-		|| string.starts_with('.')
-		|| string.starts_with("0x")
-		|| string.parse::<i64>().is_ok()
-		|| string.parse::<f64>().is_ok()
-}
-
-pub fn manifest_yaml_ex(val: &Val, options: &YamlFormat<'_>) -> Result<String> {
-	let mut out = String::new();
-	manifest_yaml_ex_buf(val, &mut out, &mut String::new(), options)?;
-	Ok(out)
-}
-
-#[allow(clippy::too_many_lines)]
-fn manifest_yaml_ex_buf(
-	val: &Val,
-	buf: &mut String,
-	cur_padding: &mut String,
-	options: &YamlFormat<'_>,
-) -> Result<()> {
-	match val {
-		Val::Bool(v) => {
-			if *v {
-				buf.push_str("true");
-			} else {
-				buf.push_str("false");
-			}
-		}
-		Val::Null => buf.push_str("null"),
-		Val::Str(s) => {
-			if s.is_empty() {
-				buf.push_str("\"\"");
-			} else if let Some(s) = s.strip_suffix('\n') {
-				buf.push('|');
-				for line in s.split('\n') {
-					buf.push('\n');
-					buf.push_str(cur_padding);
-					buf.push_str(&options.padding);
-					buf.push_str(line);
-				}
-			} else if !options.quote_keys && !yaml_needs_quotes(s) {
-				buf.push_str(s);
-			} else {
-				escape_string_json_buf(s, buf);
-			}
-		}
-		Val::Num(n) => write!(buf, "{}", *n).unwrap(),
-		Val::Arr(a) => {
-			if a.is_empty() {
-				buf.push_str("[]");
-			} else {
-				for (i, item) in a.iter().enumerate() {
-					if i != 0 {
-						buf.push('\n');
-						buf.push_str(cur_padding);
-					}
-					let item = item?;
-					buf.push('-');
-					match &item {
-						Val::Arr(a) if !a.is_empty() => {
-							buf.push('\n');
-							buf.push_str(cur_padding);
-							buf.push_str(&options.padding);
-						}
-						_ => buf.push(' '),
-					}
-					let extra_padding = match &item {
-						Val::Arr(a) => !a.is_empty(),
-						Val::Obj(o) => !o.is_empty(),
-						_ => false,
-					};
-					let prev_len = cur_padding.len();
-					if extra_padding {
-						cur_padding.push_str(&options.padding);
-					}
-					manifest_yaml_ex_buf(&item, buf, cur_padding, options)?;
-					cur_padding.truncate(prev_len);
-				}
-			}
-		}
-		Val::Obj(o) => {
-			if o.is_empty() {
-				buf.push_str("{}");
-			} else {
-				for (i, key) in o
-					.fields(
-						#[cfg(feature = "exp-preserve-order")]
-						options.preserve_order,
-					)
-					.iter()
-					.enumerate()
-				{
-					if i != 0 {
-						buf.push('\n');
-						buf.push_str(cur_padding);
-					}
-					if !options.quote_keys && !yaml_needs_quotes(key) {
-						buf.push_str(key);
-					} else {
-						escape_string_json_buf(key, buf);
-					}
-					buf.push(':');
-					let prev_len = cur_padding.len();
-					let item = o.get(key.clone())?.expect("field exists");
-					match &item {
-						Val::Arr(a) if !a.is_empty() => {
-							buf.push('\n');
-							buf.push_str(cur_padding);
-							buf.push_str(&options.arr_element_padding);
-							cur_padding.push_str(&options.arr_element_padding);
-						}
-						Val::Obj(o) if !o.is_empty() => {
-							buf.push('\n');
-							buf.push_str(cur_padding);
-							buf.push_str(&options.padding);
-							cur_padding.push_str(&options.padding);
-						}
-						_ => buf.push(' '),
-					}
-					manifest_yaml_ex_buf(&item, buf, cur_padding, options)?;
-					cur_padding.truncate(prev_len);
-				}
-			}
-		}
-		Val::Func(_) => throw!("tried to manifest function"),
-	}
-	Ok(())
 }
